fix: move local db code to another folder
